Move to `enum_utils` header
